chore: move hooks to hooks file
